requirements.txt
fastapi[all]
python-jose[cryptography]
passlib
sqlalchemy
python-dotenv
solapi